+++ db/models/index.js
@@ -2,12 +2,14 @@ const { User, UserSchema } = require('./user.model');
 const { Profile, ProfileSchema } = require('./profile.model');
 const { Auth, AuthSchema } = require('./auth.model');
 const { Post, PostSchema } = require('./post.model');
+const { Follower, FollowerSchema } =  require('./follower.model');
 
 function setupModels(sequelize) {
     User.init(UserSchema, User.config(sequelize));
     Profile.init(ProfileSchema, Profile.config(sequelize));
     Auth.init(AuthSchema, Auth.config(sequelize));
     Post.init(PostSchema, Post.config(sequelize));
+    Follower.init(FollowerSchema, Follower.config(sequelize));
 
     User.associate(sequelize.models);
     Profile.associate(sequelize.models);