+++ index.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const cors = require('cors');
+const router = require('./routes/router');
 
 const app = express();
 const port = process.env.PORT || 3000;
@@ -8,6 +9,7 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
 app.use(cors());
+app.use(router);
 
 app.listen(port, () => {
   console.log('Mi port' +  port);