+++ modules/user/serivices.js
@@ -1,0 +1,12 @@
+const { models } = require('../../services/sequelize/sequelize.service');
+
+class UserService {
+    constructor() {}
+
+    async createUser(data) {
+        const newUser = await models.User.create(data);
+        return newUser;
+    }
+}
+
+module.exports = UserService;