+++ routes/router.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const user_router = require('../modules/user/router');
+const router = express.Router();
+
+router.use('/api/v1/users', user_router);
+
+module.exports = router;